backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
websockets
httpx
python-multipart
sqlalchemy
aiosqlite
langgraph
